pyproject.toml
[tool.poetry]
name = "make_desicion"
version = "0.1.0"
description = ""
authors = ["Melvin Mendoza <melvin.mendoza@digitalreef.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
pydantic = "^1.10.2"
requests = "^2.28.1"
db-sqlite3 = "^0.0.1"
SQLAlchemy = "^1.4.44"
sqlalchemy-orm = "^1.2.3"
pysqlite3 = "^0.4.7"
pandas = "^1.5.2"
scikit-learn = "^1.1.3"
matplotlib = "^3.6.2"
fastapi = "^0.87.0"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
